Features header navigation bar
The magic occurs in header.html and navigation.html. The site.navigation
global is required to be defined in order to iterate through the
specified list items in _config.yml.
On the other hand some style has been applied. To be honest, it's just a
copy-paste from bootstraps jumbotron header with pills.